Add dynamic link to create_customer_address
This is needed to assign the address to the customer. I've assigned the doc of the new address to the variable new_address, because I think it's more clear to use new_address.name when creating the dynamic link.

dynamic_link = ... might me not necessary, but thats up to the core team to decide. What do you think?